Add ability to enter the PIN on the device itself.
[ NO  |  YES ]

A short click of yes will move the cursor to the right.
A short click of no will move the cursor up.

A medium long click (until it blinks) will select a digit for entry.

A very long click will select the current digit and finish. (4x longer than the medium click).